@@ -1,6 +1,8 @@
 ﻿namespace Nerd_STF.Mathematics.Geometry;
 
-public struct Box3D : ICloneable, IContainer<Vert>, IEquatable<Box3D>
+public record class Box3D : IAbsolute<Box3D>, IAverage<Box3D>, ICeiling<Box3D>, IClamp<Box3D>,
+    IContains<Vert>, IEquatable<Box3D>, IFloor<Box3D>, ILerp<Box3D, float>, IMedian<Box3D>,
+    IRound<Box3D>, IShape3D<float>, ISplittable<Box3D, (Vert[] centers, Float3[] sizes)>
 {
     public static Box3D Unit => new(Vert.Zero, Float3.One);
 
@@ -51,10 +53,12 @@ public struct Box3D : ICloneable, IContainer<Vert>, IEquatable<Box3D>
         (Vert[] centers, Float3[] sizes) = SplitArray(vals);
         return new(Vert.Average(centers), Float3.Average(sizes));
     }
-    public static Box3D Ceiling(Box3D val) => new(Vert.Ceiling(val.center), Float3.Ceiling(val.size));
+    public static Box3D Ceiling(Box3D val) =>
+        new(Vert.Ceiling(val.center), (Float3)Float3.Ceiling(val.size));
     public static Box3D Clamp(Box3D val, Box3D min, Box3D max) =>
         new(Vert.Clamp(val.center, min.center, max.center), Float3.Clamp(val.size, min.size, max.size));
-    public static Box3D Floor(Box3D val) => new(Vert.Floor(val.center), Float3.Floor(val.size));
+    public static Box3D Floor(Box3D val) =>
+        new(Vert.Floor(val.center), (Float3)Float3.Floor(val.size));
     public static Box3D Lerp(Box3D a, Box3D b, float t, bool clamp = true) =>
         new(Vert.Lerp(a.center, b.center, t, clamp), Float3.Lerp(a.size, b.size, t, clamp));
     public static Box3D Median(params Box3D[] vals)
@@ -62,16 +66,8 @@ public struct Box3D : ICloneable, IContainer<Vert>, IEquatable<Box3D>
         (Vert[] verts, Float3[] sizes) = SplitArray(vals);
         return new(Vert.Median(verts), Float3.Median(sizes));
     }
-    public static Box3D Max(params Box3D[] vals)
-    {
-        (Vert[] verts, Float3[] sizes) = SplitArray(vals);
-        return new(Vert.Max(verts), Float3.Max(sizes));
-    }
-    public static Box3D Min(params Box3D[] vals)
-    {
-        (Vert[] verts, Float3[] sizes) = SplitArray(vals);
-        return new(Vert.Min(verts), Float3.Min(sizes));
-    }
+    public static Box3D Round(Box3D val) => new(Vert.Ceiling(val.center), (Float3)Float3.Ceiling(val.size));
+
     public static (Vert[] centers, Float3[] sizes) SplitArray(params Box3D[] vals)
     {
         Vert[] centers = new Vert[vals.Length];
@@ -86,25 +82,27 @@ public struct Box3D : ICloneable, IContainer<Vert>, IEquatable<Box3D>
         return (centers, sizes);
     }
 
-    public override bool Equals([NotNullWhen(true)] object? obj)
+    public virtual bool Equals(Box3D? other)
     {
-        if (obj == null || obj.GetType() != typeof(Box3D)) return base.Equals(obj);
-        return Equals((Box3D)obj);
+        if (other is null) return false;
+        return center == other.center && size == other.size;
     }
-    public bool Equals(Box3D other) => center == other.center && size == other.size;
-    public override int GetHashCode() => center.GetHashCode() ^ size.GetHashCode();
-    public override string ToString() => ToString((string?)null);
-    public string ToString(string? provider) =>
-        "Min: " + MinVert.ToString(provider) + " Max: " + MaxVert.ToString(provider);
-    public string ToString(IFormatProvider provider) =>
-        "Min: " + MinVert.ToString(provider) + " Max: " + MaxVert.ToString(provider);
+    public override int GetHashCode() => base.GetHashCode();
 
     public bool Contains(Vert vert)
     {
         Float3 diff = Float3.Absolute(center - vert);
         return diff.x <= size.x && diff.y <= size.y && diff.z <= size.z;
     }
-    public object Clone() => new Box3D(center, size);
+
+    protected virtual bool PrintMembers(StringBuilder builder)
+    {
+        builder.Append("Min = ");
+        builder.Append(MinVert);
+        builder.Append(", Max = ");
+        builder.Append(MaxVert);
+        return true;
+    }
 
     public static Box3D operator +(Box3D a, Vert b) => new(a.center + b, a.size);
     public static Box3D operator +(Box3D a, Float3 b) => new(a.center, a.size + b);
@@ -115,8 +113,6 @@ public struct Box3D : ICloneable, IContainer<Vert>, IEquatable<Box3D>
     public static Box3D operator *(Box3D a, Float3 b) => new(a.center, a.size * b);
     public static Box3D operator /(Box3D a, float b) => new(a.center / b, a.size / b);
     public static Box3D operator /(Box3D a, Float3 b) => new(a.center, a.size / b);
-    public static bool operator ==(Box3D a, Box3D b) => a.Equals(b);
-    public static bool operator !=(Box3D a, Box3D b) => !a.Equals(b);
 
     public static implicit operator Box3D(Fill<float> fill) => new(fill);
     public static implicit operator Box3D(Box2D box) => new(box);
