@@ -1,6 +1,8 @@
 ﻿namespace Nerd_STF.Graphics;
 
-public struct CMYKA : IColorFloat, IEquatable<CMYKA>
+public record struct CMYKA : IAverage<CMYKA>, IClamp<CMYKA>, IColorFloat<CMYKA>, IColorPresets<CMYKA>,
+    IEquatable<CMYKA>, IIndexAll<float>, IIndexRangeAll<float>, ILerp<CMYKA, float>, IMedian<CMYKA>,
+    ISplittable<CMYKA, (float[] Cs, float[] Ms, float[] Ys, float[] Ks, float[] As)>
 {
     public static CMYKA Black => new(0, 0, 0, 1);
     public static CMYKA Blue => new(1, 1, 0, 0);
@@ -103,6 +105,28 @@ public struct CMYKA : IColorFloat, IEquatable<CMYKA>
             }
         }
     }
+    public float this[Index index]
+    {
+        get => this[index.IsFromEnd ? 5 - index.Value : index.Value];
+        set => this[index.IsFromEnd ? 5 - index.Value : index.Value] = value;
+    }
+    public float[] this[Range range]
+    {
+        get
+        {
+            int start = range.Start.IsFromEnd ? 5 - range.Start.Value : range.Start.Value;
+            int end = range.End.IsFromEnd ? 5 - range.End.Value : range.End.Value;
+            List<float> res = new();
+            for (int i = start; i < end; i++) res.Add(this[i]);
+            return res.ToArray();
+        }
+        set
+        {
+            int start = range.Start.IsFromEnd ? 5 - range.Start.Value : range.Start.Value;
+            int end = range.End.IsFromEnd ? 5 - range.End.Value : range.End.Value;
+            for (int i = start; i < end; i++) this[i] = value[i];
+        }
+    }
 
     public static CMYKA Average(params CMYKA[] vals)
     {
@@ -111,16 +135,12 @@ public struct CMYKA : IColorFloat, IEquatable<CMYKA>
 
         return val / vals.Length;
     }
-    public static CMYKA Ceiling(CMYKA val) => new(Mathf.Ceiling(val.C), Mathf.Ceiling(val.M),
-        Mathf.Ceiling(val.Y), Mathf.Ceiling(val.K), Mathf.Ceiling(val.A));
     public static CMYKA Clamp(CMYKA val, CMYKA min, CMYKA max) =>
         new(Mathf.Clamp(val.C, min.C, max.C),
             Mathf.Clamp(val.M, min.M, max.M),
             Mathf.Clamp(val.Y, min.Y, max.Y),
             Mathf.Clamp(val.K, min.K, max.K),
             Mathf.Clamp(val.A, min.A, max.A));
-    public static CMYKA Floor(CMYKA val) => new(Mathf.Floor(val.C), Mathf.Floor(val.M),
-        Mathf.Floor(val.Y), Mathf.Floor(val.K), Mathf.Floor(val.A));
     public static CMYKA Lerp(CMYKA a, CMYKA b, float t, bool clamp = true) =>
         new(Mathf.Lerp(a.C, b.C, t, clamp), Mathf.Lerp(a.M, b.M, t, clamp), Mathf.Lerp(a.Y, b.Y, t, clamp),
             Mathf.Lerp(a.K, b.K, t, clamp), Mathf.Lerp(a.A, b.A, t, clamp));
@@ -136,18 +156,6 @@ public struct CMYKA : IColorFloat, IEquatable<CMYKA>
         CMYKA valA = vals[Mathf.Floor(index)], valB = vals[Mathf.Ceiling(index)];
         return Average(valA, valB);
     }
-    public static CMYKA Max(params CMYKA[] vals)
-    {
-        (float[] Cs, float[] Ms, float[] Ys, float[] Ks, float[] As) = SplitArray(vals);
-        return new(Mathf.Max(Cs), Mathf.Max(Ms), Mathf.Max(Ys), Mathf.Max(Ks), Mathf.Max(As));
-    }
-    public static CMYKA Min(params CMYKA[] vals)
-    {
-        (float[] Cs, float[] Ms, float[] Ys, float[] Ks, float[] As) = SplitArray(vals);
-        return new(Mathf.Min(Cs), Mathf.Min(Ms), Mathf.Min(Ys), Mathf.Min(Ks), Mathf.Min(As));
-    }
-    public static CMYKA Round(CMYKA val) => new(Mathf.Round(val.C), Mathf.Round(val.M),
-        Mathf.Round(val.Y), Mathf.Round(val.K), Mathf.Round(val.A));
 
     public static (float[] Cs, float[] Ms, float[] Ys, float[] Ks, float[] As) SplitArray(params CMYKA[] vals)
     {
@@ -165,33 +173,10 @@ public struct CMYKA : IColorFloat, IEquatable<CMYKA>
         return (Cs, Ms, Ys, Ks, As);
     }
 
-    public bool Equals(IColorFloat? col) => col != null && Equals(col.ToCMYKA());
-    public bool Equals(IColorByte? col) => col != null && Equals(col.ToCMYKA());
+    public bool Equals(IColor? col) => col != null && Equals(col.ToCMYKA());
     public bool Equals(CMYKA col) => A == 0 && col.A == 0 || K == 1 && col.K == 1 || C == col.C && M == col.M
         && Y == col.Y && K == col.K && A == col.A;
-    public override bool Equals([NotNullWhen(true)] object? obj)
-    {
-        if (obj == null) return base.Equals(obj);
-        Type t = obj.GetType();
-        if (t == typeof(CMYKA)) return Equals((CMYKA)obj);
-        else if (t == typeof(RGBA)) return Equals((IColorFloat)obj);
-        else if (t == typeof(HSVA)) return Equals((IColorFloat)obj);
-        else if (t == typeof(IColorFloat)) return Equals((IColorFloat)obj);
-        else if (t == typeof(RGBAByte)) return Equals((IColorByte)obj);
-        else if (t == typeof(CMYKAByte)) return Equals((IColorByte)obj);
-        else if (t == typeof(HSVAByte)) return Equals((IColorByte)obj);
-        else if (t == typeof(IColorByte)) return Equals((IColorByte)obj);
-
-        return base.Equals(obj);
-    }
-    public override int GetHashCode() => C.GetHashCode() ^ M.GetHashCode() ^ Y.GetHashCode() ^ K.GetHashCode() ^ A.GetHashCode();
-    public string ToString(IFormatProvider provider) => "C: " + C.ToString(provider) + " M: " + M.ToString(provider)
-        + " Y: " + Y.ToString(provider) + " K: " + K.ToString(provider)
-        + " A: " + A.ToString(provider);
-    public string ToString(string? provider) => "C: " + C.ToString(provider) + " M: " + M.ToString(provider)
-        + " Y: " + Y.ToString(provider) + " K: " + K.ToString(provider)
-        + " A: " + A.ToString(provider);
-    public override string ToString() => ToString((string?)null);
+    public override int GetHashCode() => base.GetHashCode();
 
     public RGBA ToRGBA()
     {
@@ -224,7 +209,20 @@ public struct CMYKA : IColorFloat, IEquatable<CMYKA>
         yield return A;
     }
 
-    public object Clone() => new CMYKA(C, M, Y, K, A);
+    private bool PrintMembers(StringBuilder builder)
+    {
+        builder.Append("C = ");
+        builder.Append(C);
+        builder.Append(", M = ");
+        builder.Append(M);
+        builder.Append(", Y = ");
+        builder.Append(Y);
+        builder.Append(", K = ");
+        builder.Append(K);
+        builder.Append(", A = ");
+        builder.Append(A);
+        return true;
+    }
 
     public static CMYKA operator +(CMYKA a, CMYKA b) => new(a.C + b.C, a.M + b.M, a.Y + b.Y, a.K + b.K, a.A + b.A);
     public static CMYKA operator -(CMYKA c) => new(1 - c.C, 1 - c.M, 1 - c.Y, 1 - c.K, c.A != 1 ? 1 - c.A : 1);
@@ -233,25 +231,23 @@ public struct CMYKA : IColorFloat, IEquatable<CMYKA>
     public static CMYKA operator *(CMYKA a, float b) => new(a.C * b, a.M * b, a.Y * b, a.K * b, a.A * b);
     public static CMYKA operator /(CMYKA a, CMYKA b) => new(a.C / b.C, a.M / b.M, a.Y / b.Y, a.K / b.K, a.A / b.A);
     public static CMYKA operator /(CMYKA a, float b) => new(a.C / b, a.M / b, a.Y / b, a.K / b, a.A / b);
-    public static bool operator ==(CMYKA a, RGBA b) => a.Equals(b);
-    public static bool operator !=(CMYKA a, RGBA b) => !a.Equals(b);
-    public static bool operator ==(CMYKA a, CMYKA b) => a.Equals(b);
-    public static bool operator !=(CMYKA a, CMYKA b) => !a.Equals(b);
     public static bool operator ==(CMYKA a, HSVA b) => a.Equals(b);
-    public static bool operator !=(CMYKA a, HSVA b) => !a.Equals(b);
-    public static bool operator ==(CMYKA a, RGBAByte b) => a.Equals((IColorByte?)b);
-    public static bool operator !=(CMYKA a, RGBAByte b) => !a.Equals((IColorByte?)b);
-    public static bool operator ==(CMYKA a, CMYKAByte b) => a.Equals((IColorByte?)b);
-    public static bool operator !=(CMYKA a, CMYKAByte b) => !a.Equals((IColorByte?)b);
-    public static bool operator ==(CMYKA a, HSVAByte b) => a.Equals((IColorByte?)b);
-    public static bool operator !=(CMYKA a, HSVAByte b) => !a.Equals((IColorByte?)b);
+    public static bool operator !=(CMYKA a, HSVA b) => a.Equals(b);
+    public static bool operator ==(CMYKA a, RGBA b) => a.Equals(b);
+    public static bool operator !=(CMYKA a, RGBA b) => a.Equals(b);
+    public static bool operator ==(CMYKA a, CMYKAByte b) => a.Equals(b);
+    public static bool operator !=(CMYKA a, CMYKAByte b) => a.Equals(b);
+    public static bool operator ==(CMYKA a, HSVAByte b) => a.Equals(b);
+    public static bool operator !=(CMYKA a, HSVAByte b) => a.Equals(b);
+    public static bool operator ==(CMYKA a, RGBAByte b) => a.Equals(b);
+    public static bool operator !=(CMYKA a, RGBAByte b) => a.Equals(b);
 
     public static explicit operator CMYKA(Float3 val) => new(val.x, val.y, val.z, 0);
     public static implicit operator CMYKA(Float4 val) => new(val.x, val.y, val.z, val.w);
-    public static implicit operator CMYKA(RGBA val) => val.ToCMYKA();
     public static implicit operator CMYKA(HSVA val) => val.ToCMYKA();
-    public static implicit operator CMYKA(RGBAByte val) => val.ToCMYKA();
+    public static implicit operator CMYKA(RGBA val) => val.ToCMYKA();
     public static implicit operator CMYKA(CMYKAByte val) => val.ToCMYKA();
     public static implicit operator CMYKA(HSVAByte val) => val.ToCMYKA();
+    public static implicit operator CMYKA(RGBAByte val) => val.ToCMYKA();
     public static implicit operator CMYKA(Fill<float> val) => new(val);
 }
