@@ -1,6 +1,14 @@
-﻿namespace Nerd_STF.Mathematics;
+﻿using System;
 
-public struct Float3 : ICloneable, IComparable<Float3>, IEquatable<Float3>, IGroup<float>
+namespace Nerd_STF.Mathematics;
+
+public record struct Float3 : IAbsolute<Float3>, IAverage<Float3>,
+    ICeiling<Float3, Int3>, IClamp<Float3>, IClampMagnitude<Float3, float>, IComparable<Float3>,
+    ICross<Float3>, IDivide<Float3>, IDot<Float3, float>, IEquatable<Float3>,
+    IFloor<Float3, Int3>, IFromTuple<Float3, (float x, float y, float z)>, IGroup<float>,
+    IIndexAll<float>, IIndexRangeAll<float>, ILerp<Float3, float>, IMathOperators<Float3>, IMax<Float3>,
+    IMedian<Float3>, IMin<Float3>, IPresets3D<Float3>, IProduct<Float3>, IRound<Float3, Int3>,
+    ISplittable<Float3, (float[] Xs, float[] Ys, float[] Zs)>, ISubtract<Float3>, ISum<Float3>
 {
     public static Float3 Back => new(0, 0, -1);
     public static Float3 Down => new(0, -1, 0);
@@ -61,11 +69,33 @@ public struct Float3 : ICloneable, IComparable<Float3>, IEquatable<Float3>, IGro
             }
         }
     }
+    public float this[Index index]
+    {
+        get => this[index.IsFromEnd ? 3 - index.Value : index.Value];
+        set => this[index.IsFromEnd ? 3 - index.Value : index.Value] = value;
+    }
+    public float[] this[Range range]
+    {
+        get
+        {
+            int start = range.Start.IsFromEnd ? 3 - range.Start.Value : range.Start.Value;
+            int end = range.End.IsFromEnd ? 3 - range.End.Value : range.End.Value;
+            List<float> res = new();
+            for (int i = start; i < end; i++) res.Add(this[i]);
+            return res.ToArray();
+        }
+        set
+        {
+            int start = range.Start.IsFromEnd ? 3 - range.Start.Value : range.Start.Value;
+            int end = range.End.IsFromEnd ? 3 - range.End.Value : range.End.Value;
+            for (int i = start; i < end; i++) this[i] = value[i];
+        }
+    }
 
     public static Float3 Absolute(Float3 val) =>
         new(Mathf.Absolute(val.x), Mathf.Absolute(val.y), Mathf.Absolute(val.z));
     public static Float3 Average(params Float3[] vals) => Sum(vals) / vals.Length;
-    public static Float3 Ceiling(Float3 val) =>
+    public static Int3 Ceiling(Float3 val) =>
         new(Mathf.Ceiling(val.x), Mathf.Ceiling(val.y), Mathf.Ceiling(val.z));
     public static Float3 Clamp(Float3 val, Float3 min, Float3 max) =>
         new(Mathf.Clamp(val.x, min.x, max.x),
@@ -103,7 +133,7 @@ public struct Float3 : ICloneable, IComparable<Float3>, IEquatable<Float3>, IGro
         }
         return x + y + z;
     }
-    public static Float3 Floor(Float3 val) =>
+    public static Int3 Floor(Float3 val) =>
         new(Mathf.Floor(val.x), Mathf.Floor(val.y), Mathf.Floor(val.z));
     public static Float3 Lerp(Float3 a, Float3 b, float t, bool clamp = true) =>
         new(Mathf.Lerp(a.x, b.x, t, clamp), Mathf.Lerp(a.y, b.y, t, clamp), Mathf.Lerp(a.z, b.z, t, clamp));
@@ -117,14 +147,14 @@ public struct Float3 : ICloneable, IComparable<Float3>, IEquatable<Float3>, IGro
     {
         if (vals.Length < 1) return Zero;
         Float3 val = vals[0];
-        foreach (Float3 d in vals) val = d > val ? d : val;
+        foreach (Float3 d in vals) val = d.Magnitude > val.Magnitude ? d : val;
         return val;
     }
     public static Float3 Min(params Float3[] vals)
     {
         if (vals.Length < 1) return Zero;
         Float3 val = vals[0];
-        foreach (Float3 d in vals) val = d < val ? d : val;
+        foreach (Float3 d in vals) val = d.Magnitude < val.Magnitude ? d : val;
         return val;
     }
     public static Float3 Product(params Float3[] vals)
@@ -134,8 +164,8 @@ public struct Float3 : ICloneable, IComparable<Float3>, IEquatable<Float3>, IGro
         foreach (Float3 d in vals) val *= d;
         return val;
     }
-    public static Float3 Round(Float3 val) =>
-        new(Mathf.Round(val.x), Mathf.Round(val.y), Mathf.Round(val.z));
+    public static Int3 Round(Float3 val) =>
+        new(Mathf.RoundInt(val.x), Mathf.RoundInt(val.y), Mathf.RoundInt(val.z));
     public static Float3 Subtract(Float3 num, params Float3[] vals) => num - Sum(vals);
     public static Float3 Sum(params Float3[] vals)
     {
@@ -156,21 +186,11 @@ public struct Float3 : ICloneable, IComparable<Float3>, IEquatable<Float3>, IGro
         return (Xs, Ys, Zs);
     }
 
+    [Obsolete("This method is a bit ambiguous. You should instead compare " +
+        nameof(Magnitude) + "s directly.")]
     public int CompareTo(Float3 other) => Magnitude.CompareTo(other.Magnitude);
-    public override bool Equals([NotNullWhen(true)] object? obj)
-    {
-        if (obj == null || obj.GetType() != typeof(Float3)) return base.Equals(obj);
-        return Equals((Float3)obj);
-    }
     public bool Equals(Float3 other) => x == other.x && y == other.y && z == other.z;
-    public override int GetHashCode() => x.GetHashCode() ^ y.GetHashCode() ^ z.GetHashCode();
-    public override string ToString() => ToString((string?)null);
-    public string ToString(string? provider) =>
-        "X: " + x.ToString(provider) + " Y: " + y.ToString(provider) + " Z: " + z.ToString(provider);
-    public string ToString(IFormatProvider provider) =>
-        "X: " + x.ToString(provider) + " Y: " + y.ToString(provider) + " Z: " + z.ToString(provider);
-
-    public object Clone() => new Float3(x, y, z);
+    public override int GetHashCode() => base.GetHashCode();
 
     IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
     public IEnumerator<float> GetEnumerator()
@@ -197,6 +217,17 @@ public struct Float3 : ICloneable, IComparable<Float3>, IEquatable<Float3>, IGro
         return new(yaw, pitch, mag);
     }
 
+    private bool PrintMembers(StringBuilder builder)
+    {
+        builder.Append("x = ");
+        builder.Append(x);
+        builder.Append(", y = ");
+        builder.Append(y);
+        builder.Append(", z = ");
+        builder.Append(z);
+        return true;
+    }
+
     public static Float3 operator +(Float3 a, Float3 b) => new(a.x + b.x, a.y + b.y, a.z + b.z);
     public static Float3 operator -(Float3 d) => new(-d.x, -d.y, -d.z);
     public static Float3 operator -(Float3 a, Float3 b) => new(a.x - b.x, a.y - b.y, a.z - b.z);
@@ -207,11 +238,17 @@ public struct Float3 : ICloneable, IComparable<Float3>, IEquatable<Float3>, IGro
     public static Float3 operator /(Float3 a, Float3 b) => new(a.x / b.x, a.y / b.y, a.z / b.z);
     public static Float3 operator /(Float3 a, float b) => new(a.x / b, a.y / b, a.z / b);
     public static Float3 operator /(Float3 a, Matrix b) => (Float3)((Matrix)a / b);
-    public static bool operator ==(Float3 a, Float3 b) => a.Equals(b);
-    public static bool operator !=(Float3 a, Float3 b) => !a.Equals(b);
+    [Obsolete("This operator is a bit ambiguous. You should instead compare " +
+        nameof(Magnitude) + "s directly.")]
     public static bool operator >(Float3 a, Float3 b) => a.CompareTo(b) > 0;
+    [Obsolete("This operator is a bit ambiguous. You should instead compare " +
+        nameof(Magnitude) + "s directly.")]
     public static bool operator <(Float3 a, Float3 b) => a.CompareTo(b) < 0;
+    [Obsolete("This operator is a bit ambiguous (and misleading at times). " +
+        "You should instead compare " + nameof(Magnitude) + "s directly.")]
     public static bool operator >=(Float3 a, Float3 b) => a == b || a > b;
+    [Obsolete("This operator is a bit ambiguous (and misleading at times). " +
+        "You should instead compare " + nameof(Magnitude) + "s directly.")]
     public static bool operator <=(Float3 a, Float3 b) => a == b || a < b;
 
     public static implicit operator Float3(Complex val) => new(val.u, val.i, 0);
@@ -230,4 +267,6 @@ public struct Float3 : ICloneable, IComparable<Float3>, IEquatable<Float3>, IGro
     public static explicit operator Float3(HSVAByte val) => (Float3)val.ToHSVA();
     public static implicit operator Float3(Fill<float> fill) => new(fill);
     public static implicit operator Float3(Fill<int> fill) => new(fill);
+    public static implicit operator Float3((float x, float y, float z) val) =>
+        new(val.x, val.y, val.z);
 }
