@@ -1,18 +1,22 @@
-﻿global using System;
+﻿global using Nerd_STF;
+global using Nerd_STF.Graphics;
+global using Nerd_STF.Graphics.Abstract;
+global using Nerd_STF.Exceptions;
+global using Nerd_STF.Extensions;
+global using Nerd_STF.Mathematics;
+global using Nerd_STF.Mathematics.Abstract;
+global using Nerd_STF.Mathematics.Algebra;
+global using Nerd_STF.Mathematics.Geometry;
+global using Nerd_STF.Mathematics.NumberSystems;
+global using Nerd_STF.Mathematics.Samples;
+global using System;
 global using System.Collections;
 global using System.Collections.Generic;
 global using System.Diagnostics.CodeAnalysis;
 global using System.IO;
 global using System.Linq;
 global using System.Net.Http;
+global using System.Runtime.Serialization;
+global using System.Text;
 global using System.Threading;
 global using System.Threading.Tasks;
-global using Nerd_STF;
-global using Nerd_STF.Graphics;
-global using Nerd_STF.Exceptions;
-global using Nerd_STF.Extensions;
-global using Nerd_STF.Mathematics;
-global using Nerd_STF.Mathematics.Algebra;
-global using Nerd_STF.Mathematics.Geometry;
-global using Nerd_STF.Mathematics.NumberSystems;
-global using Nerd_STF.Mathematics.Samples;
